requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-core
langchain-openai
langchain-chroma
langchain-community
pydantic
pypdf
